Changing the pairing process for onboarding.
* The pairing code is repeated once every 60 seconds (which is about 30 seconds of quiet in between for English)
* Text about pairing is more friendly to first-time users.

One less-than-ideal piece of this is the comment in pairing.pairing.code.dialog.  It refers to the "button", which is device specific.  Once Enclosures are abstracted well, this skill should ask the Enclosure for an additional comment to speak at the end of the pairing rather than this hard-coded value.