Add lock around cancelling of Timer
==== Tech Notes ====
Timer cancel basically only works while the Timer is waiting to be run.
During execution the request to cancel has no effect and since the Timer is
recreated at the end of execution an orphaned Timer loop can be created
at shutdown unless extra logic is added to make sure this doesn't occur.