Only inform user of capabilities after pairing sequence
If the skill was started while network is down pairing will return as
false and the handler for mycroft.ready will be registered and the "I'm
now ready for use..." message would be spoken.

This change modifies the handler to require a pairing sequence to have
been completed before uttering the message.