Fix pairing code not displaying on Mark 1
The pairing code was not displaying on the Mark 1 following the
latest update. I don't yet know why - but if you set the wait flag
on this dialog to True then it works as expected.

Setting a long sleep immediately before the dialog is spoken does
not have the same result so it is something to do with this method
or _attempt_activation()